requirements.txt
ultralytics==8.3.34
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0.76
albumentations>=1.3.1
numpy>=1.24.0
pyyaml>=6.0.1
matplotlib>=3.7.0
tensorboard>=2.14.0
onnx>=1.15.0
onnxruntime>=1.16.0
pycocotools>=2.0.7
pytest>=7.4.0
Pillow>=10.0.0
scipy>=1.10.0
rich>=13.6.0
tk>=0.1.0

